+++ client/components/navbar.js
@@ -1,30 +1,63 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
-import {Link} from 'react-router-dom'
+import {Link, NavLink} from 'react-router-dom'
 import {logout} from '../store'
 
-const Navbar = ({handleClick, isLoggedIn}) => (
+const Navbar = ({handleClick, isLoggedIn, isAdmin}) => (
   <div>
-    <h1>BOILERMAKER</h1>
-    <nav>
+    <nav className="navbar">
       {isLoggedIn ? (
         <div>
-          {/* The navbar will show these links after you log in */}
-          <Link to="/home">Home</Link>
-          <a href="#" onClick={handleClick}>
-            Logout
-          </a>
+          <ul className="nav-links">
+            <li className="nav-item">
+              <NavLink id="shirt-logo" to="/">
+                <h3>lets trend..</h3>
+              </NavLink>
+            </li>
+            <li className="nav-item">
+              <NavLink to="/profile">
+                <h5>Profile</h5>
+              </NavLink>
+            </li>
+            <li className="nav-item">
+              <NavLink to="#" onClick={handleClick}>
+                <h5>Logout</h5>
+              </NavLink>
+            </li>
+            {isAdmin ? (
+              <li className="nav-item">
+                <NavLink to="/allUsers">
+                  <h3>All Users</h3>
+                </NavLink>
+              </li>
+            ) : (
+              ''
+            )}
+          </ul>
         </div>
       ) : (
         <div>
-          {/* The navbar will show these links before you log in */}
-          <Link to="/login">Login</Link>
-          <Link to="/signup">Sign Up</Link>
+          <ul className="nav-links">
+            <li className="nav-item">
+              <NavLink id="shirt-logo" to="/">
+                <h3>lets trend...</h3>
+              </NavLink>
+            </li>
+            <li className="nav-item">
+              <NavLink to="/login">
+                <h3>Login</h3>
+              </NavLink>
+            </li>
+            <li className="nav-item">
+              <NavLink to="/signup">
+                <h3>Sign Up</h3>
+              </NavLink>
+            </li>
+          </ul>
         </div>
       )}
     </nav>
-    <hr />
   </div>
 )
 
@@ -33,7 +66,8 @@ const Navbar = ({handleClick, isLoggedIn}) => (
  */
 const mapState = state => {
   return {
-    isLoggedIn: !!state.user.id
+    isLoggedIn: !!state.user.id,
+    isAdmin: state.user.isAdmin
   }
 }