+++ client/index.js
@@ -6,9 +6,6 @@ import history from './history'
 import store from './store'
 import App from './app'
 
-// establishes socket connection
-import './socket'
-
 ReactDOM.render(
   <Provider store={store}>
     <Router history={history}>