+++ CHANGELOG.md
@@ -55,3 +55,7 @@
     - Found out that as long as react- is 16.4+, the updates should be fine
 - react update to 16.8.6 from 16.4.2
 - react-dom update to 16.8.6 from 16.4.2
+
+## Thursday, April 11th, 2019
+### Dependencies 
+- sequelize update to 5.3.1 from 5.2.15